requirements.txt
Django==5.2.13
djangorestframework==3.17.1
django-cors-headers==4.3.1